+++ tests/unit/models/test__utils.py
@@ -1,0 +1,80 @@
+"""Test the ``_utils`` module."""
+
+import re
+import sys
+from unittest.mock import patch
+
+import pytest
+import torch
+
+from deepecho.models._utils import _set_device, _validate_gpu_parameter, validate_and_set_device
+
+
+def test__validate_gpu_parameter():
+    """Test the ``_validate_gpu_parameter`` method."""
+    # Setup
+    expected_error = re.escape(
+        'Cannot resolve the provided values of `cuda` and `enable_gpu` parameters. '
+        'Please use only `enable_gpu`.'
+    )
+    expected_warning = re.escape(
+        '`cuda` parameter is deprecated and will be removed in a future release. '
+        'Please use `enable_gpu` instead.'
+    )
+
+    # Run
+    enable_gpu_1 = _validate_gpu_parameter(False, None)
+    enable_gpu_2 = _validate_gpu_parameter(True, None)
+    with pytest.warns(FutureWarning, match=expected_warning):
+        enable_gpu_3 = _validate_gpu_parameter(True, False)
+
+    with pytest.raises(ValueError, match=expected_error):
+        _validate_gpu_parameter(False, True)
+
+    # Assert
+    assert enable_gpu_1 is False
+    assert enable_gpu_2 is True
+    assert enable_gpu_3 is False
+
+
+def test__set_device():
+    """Test the ``_set_device`` method."""
+    # Run
+    device_1 = _set_device(False)
+    device_2 = _set_device(True)
+    device_3 = _set_device(True, 'cpu')
+    device_4 = _set_device(enable_gpu=False, device='cpu')
+
+    # Assert
+    if (
+        sys.platform == 'darwin'
+        and getattr(torch.backends, 'mps', None)
+        and torch.backends.mps.is_available()
+    ):
+        expected_device_2 = torch.device('mps')
+    elif torch.cuda.is_available():
+        expected_device_2 = torch.device('cuda')
+    else:
+        expected_device_2 = torch.device('cpu')
+
+    assert device_1 == torch.device('cpu')
+    assert device_2 == expected_device_2
+    assert device_3 == torch.device('cpu')
+    assert device_4 == torch.device('cpu')
+
+
+@patch('deepecho.models._utils._set_device')
+@patch('deepecho.models._utils._validate_gpu_parameter')
+def test_validate_and_set_device(mock_validate, mock_set_device):
+    """Test the ``validate_and_set_device`` method."""
+    # Setup
+    mock_validate.return_value = True
+    mock_set_device.return_value = torch.device('cuda')
+
+    # Run
+    device = validate_and_set_device(enable_gpu=True, cuda=None)
+
+    # Assert
+    mock_validate.assert_called_once_with(True, None)
+    mock_set_device.assert_called_once_with(True)
+    assert device == torch.device('cuda')
